INF-2266: Add KAFKA_DATA_DIRS to kafka environment
There's a bug in the image that will check the writability of the
directories in the environment variable KAFKA_DATA_DIRS (and
interestingly, does _not_ check the writability of KAFKA_LOG_DIRS). The
image will default KAFKA_DATA_DIRS to `/var/lib/kafka/data` if not set,
which is not a writable directory when running with a read-only root
filesystem unless a volume is mounted there. Since KAFKA_DATA_DIRS is
ignored beyond the initial writability check, it seems just adding that
value to the environment with the appropriate directories is the best
way to go.